chore: rename folder to match component
